Turn `config` into a package
