+++ reapy/config/__init__.py
@@ -1,0 +1,2 @@
+from .config import *
+from .python_sharedlib import get_python_sharedlib
